refactor(AppHeader): remove unnecessary component
