+++ components/Ui/AppHeader/index.vue
@@ -1,16 +1,88 @@
 <template>
   <header>
-    <div class="lg:hidden">
-      <UiAppHeaderMobile :logo-color="props.logoColor" />
+    <div class="container">
+      <div class="container-sub">
+        <NuxtLink :to="localePath('/')">
+          <img width="75" height="75" src="/logo.svg" />
+        </NuxtLink>
+
+        <div class="link">
+          <div class="link-nav">
+            <NuxtLink
+              class="link-nav-item"
+              :to="localePath('/') + '#statistics'"
+              >{{ $t('link.statistics') }}</NuxtLink
+            >
+            &VerticalBar;
+            <NuxtLink
+              class="link-nav-item"
+              :to="localePath('/') + '#learn-more'"
+              >{{ $t('link.learnMore') }}</NuxtLink
+            >
+          </div>
+
+          <NuxtLink
+            v-for="localeItem in availableLocales"
+            :key="localeItem.code"
+            :to="switchLocalePath(localeItem.code)"
+            class="link-locale"
+            :class="localeItem.code"
+            >{{ localeItem.name }}
+          </NuxtLink>
+
+          <div>
+            <UserMenuButton
+              @update:showItems="(value: boolean) => {showMenu = value}"
+            />
+            <UserMenuItems v-if="showMenu" />
+          </div>
+        </div>
+      </div>
+      <div class="header-separator"></div>
     </div>
-    <div class="hidden lg:block"><UiAppHeaderDesktop /></div>
   </header>
 </template>
-<script setup>
-  const props = defineProps({
-    logoColor: {
-      type: String,
-      default: 'default',
-    },
+<script setup lang="ts">
+  import { LocaleObject } from '@nuxtjs/i18n/dist/runtime/composables'
+
+  // language switcher
+  const { locale, locales } = useI18n()
+  const switchLocalePath = useSwitchLocalePath()
+  const localePath = useLocalePath()
+  const availableLocales = computed(() => {
+    return (locales.value as LocaleObject[]).filter(
+      (l: LocaleObject) => l.code !== locale.value
+    )
   })
+
+  const showMenu = ref<boolean>(false)
 </script>
+<style lang="postcss" scoped>
+  .container {
+    @apply py-2.5 pt-10 text-nuha-fushia;
+
+    &-sub {
+      @apply flex justify-between;
+    }
+  }
+
+  .link {
+    @apply flex justify-between items-center;
+
+    &-locale {
+      @apply rounded-xl border p-2 text-xl px-3.5 border-nuha-fushia flex items-center;
+    }
+
+    &-nav {
+      @apply text-xl font-light px-7;
+
+      &-item {
+        @apply px-5;
+      }
+    }
+  }
+
+  .header-separator {
+    @apply block my-2.5 border-t border-t-nuha-fushia-300 h-px w-full;
+  }
+</style>
